+++ changelog.md
@@ -3,6 +3,7 @@
 - Add basic support for json clients
 - Add data_point_path event
 - Add getDeviceDescription, getParamsetDescription, listDevices, getValue, setValue, getParamset, putParamset to json_rpc
+- Add option to refresh data by interface
 - Add xml_rpc support flag to client
 - Extend DP_KEY with interface_id
 - Rename event to data_point_event